requirements.txt
ertifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
Django==2.1
djangorestframework==3.9.4
Flask==1.1.1
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.7
pkg-resources==0.0.0
python-dateutil==2.6.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
whitenoise==4.1.2
